GH-9 - moved MainLayout to components folder as it belongs there
